# Replace SortedSet.TailSet with separate CronField container