--- a/src/Quartz/Util/SortedSetExtensions.cs
+++ b/src/Quartz/Util/SortedSetExtensions.cs
@@ -1,10 +1,36 @@
-﻿namespace Quartz.Util
+﻿namespace Quartz.Util;
+
+internal static class SortedSetExtensions
 {
-    internal static class SortedSetExtensions
+    internal static bool TryGetMinValueStartingFrom(this SortedSet<int> set, int start, out int min)
     {
-        internal static SortedSet<int> TailSet(this SortedSet<int> set, int value)
+        min = set.Min;
+
+        if (set.Contains(CronExpressionConstants.AllSpec) || set.Contains(start))
         {
-            return set.GetViewBetween(value, 9999999);
+            min = start;
+            return true;
         }
-   }
+
+        if (set.Count == 0 || set.Max < start)
+        {
+            return false;
+        }
+
+        if (set.Min >= start)
+        {
+            // value is contained and would be returned from view
+            return true;
+        }
+
+        // slow path
+        var view = set.GetViewBetween(start, int.MaxValue);
+        if (view.Count > 0)
+        {
+            min = view.Min;
+            return true;
+        }
+
+        return false;
+    }
 }