* don't store individual values when AllSpec in use
* remove AllSpecInt and NoSpecInt

## Quartz.Benchmark.CronExpressionBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |CreateExpressionsAndCalculateFireTimeAfter|17.58 μs|0.223 μs|18.75 KB|
| **New** |	| **5.769 μs (-67%)** | **0.0163 μs** | **4.73 KB (-75%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|16.37 μs|0.099 μs|17.51 KB|
| **New** |	| **5.519 μs (-66%)** | **0.0603 μs** | **5.45 KB (-69%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|16.88 μs|0.107 μs|18.7 KB|
| **New** |	| **6.242 μs (-63%)** | **0.0120 μs** | **4.45 KB (-76%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|17.97 μs|0.312 μs|20.28 KB|
| **New** |	| **6.054 μs (-66%)** | **0.0081 μs** | **3.24 KB (-84%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|18.76 μs|0.062 μs|20.8 KB|
| **New** |	| **6.598 μs (-65%)** | **0.0091 μs** | **4.16 KB (-80%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|18.33 μs|0.049 μs|20.8 KB|
| **New** |	| **6.552 μs (-64%)** | **0.0258 μs** | **4.16 KB (-80%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|18.29 μs|0.071 μs|20.4 KB|
| **New** |	| **6.319 μs (-65%)** | **0.0141 μs** | **3.43 KB (-83%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|18.07 μs|0.087 μs|20.52 KB|
| **New** |	| **6.380 μs (-65%)** | **0.0187 μs** | **3.55 KB (-83%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|18.81 μs|0.116 μs|20.81 KB|
| **New** |	| **6.746 μs (-64%)** | **0.0125 μs** | **3.84 KB (-82%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|18.42 μs|0.204 μs|20.69 KB|
| **New** |	| **6.493 μs (-65%)** | **0.0151 μs** | **4.04 KB (-80%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|18.42 μs|0.099 μs|20.78 KB|
| **New** |	| **6.914 μs (-62%)** | **0.0178 μs** | **4.12 KB (-80%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|18.36 μs|0.068 μs|20.79 KB|
| **New** |	| **7.151 μs (-61%)** | **0.0562 μs** | **4.13 KB (-80%)** |
| Old |CreateExpressionsAndCalculateFireTimeAfter|17.84 μs|0.066 μs|18.34 KB|
| **New** |	| **5.883 μs (-67%)** | **0.0155 μs** | **2.46 KB (-87%)** |


/cc @jafin